refactor: FSD 아키텍처에 맞게 docs widgets 구조 개선
