+++ apps/client/src/widgets/docs/index.ts
@@ -1,3 +1,4 @@
 export { default as DocsSidebar } from './ui/DocsSidebar';
 export { default as CodeTabs, CodeTab } from './ui/CodeTabs';
-export * from './config/docsSections';
+export * from './model/constants';
+export * from './model/types';
