+++ apps/client/src/widgets/docs/model/types.ts
@@ -1,0 +1,13 @@
+export interface DocsSection {
+  label: string;
+  href: string;
+  icon: React.ElementType;
+  children?: {
+    label: string;
+    href: string;
+    children?: {
+      label: string;
+      href: string;
+    }[];
+  }[];
+}
